+++ src/main/java/redstonedubstep/mods/vanishmod/VanishUtil.java
@@ -58,7 +58,12 @@ public class VanishUtil {
 	}
 
 	public static void updateVanishedStatus(ServerPlayerEntity player, boolean vanished) {
-		player.getPersistentData().getCompound(PlayerEntity.PERSISTED_NBT_TAG).putBoolean("Vanished", vanished);
+		CompoundNBT persistentData = player.getPersistentData();
+		CompoundNBT deathPersistentData = persistentData.getCompound(PlayerEntity.PERSISTED_NBT_TAG);
+
+		deathPersistentData.putBoolean("Vanished", vanished);
+		persistentData.put(PlayerEntity.PERSISTED_NBT_TAG, deathPersistentData); //Because the deathPersistentData could have been created newly if it didn't exist before
+
 		player.setInvisible(vanished);
 
 		if (ModList.get().isLoaded("minecraft2discord")) {