+++ Changelog.md
@@ -1,3 +1,7 @@
+-----------Changelog for v1.0.9-----------
+
+- Fixed that players could not get vanished in v1.0.8 if they didn't have any persistent player data stored
+
 -----------Changelog for v1.0.8-----------
 
 - Improved mod compatibility by marking vanished players as spectators for other mods